請求入力エリアの処理更新
